backend/requirements.txt
pytesseract
requests
pillow
fastapi
uvicorn[standard]
pdf2image
pydantic
sqlalchemy
psycopg2